+++ packages/vscode/src/const.ts
@@ -5,7 +5,7 @@ export const extensionTitle = 'LikeC4'
 
 export const fileExtensions = ['.c4', '.likec4', '.like-c4'] as const
 
-export const globPattern = '**/*.{c4,likec4,like-c4}'
+export const globPattern = '**/?*.{c4,likec4,like-c4}'
 
 export const isWebUi = () => vscode.env.uiKind === vscode.UIKind.Web
 export const isVirtual = () => vscode.workspace.workspaceFolders?.every(f => f.uri.scheme !== 'file') || false